packages/neons-bidder/requirements.txt
web3==5.23.0
python-decouple==3.7